Rename some camelCase files to kebab-case
