+++ app/(main)/[slug]/[[...variant]]/page.tsx
@@ -4,8 +4,8 @@ import pMap from 'p-map'
 import type { Pokemon, PokemonForm, PokemonSpecies } from 'pokedex-promise-v2'
 import pokeapi from '@/lib/api/pokeapi'
 import { getTestSpeciesList } from '@/lib/providers'
-import { getTranslation, TypeKey, TypeLabels } from '@/lib/utils/pokeapiHelpers'
-import { excludedForms, excludedVariants } from '@/lib/utils/excludedSlugs'
+import { getTranslation, TypeKey, TypeLabels } from '@/lib/utils/pokeapi-helpers'
+import { excludedForms, excludedVariants } from '@/lib/utils/excluded-slugs'
 import LoadingSection from '@/components/details/LoadingSection'
 import LoadingMetadata from '@/components/details/LoadingMetadata'
 import MonsterMetadata from '@/components/details/MonsterMetadata'
