+++ app/(main)/page.tsx
@@ -3,8 +3,8 @@ import pMap from 'p-map'
 import { Pokemon, PokemonSpecies } from 'pokedex-promise-v2'
 import pokeapi from '@/lib/api/pokeapi'
 import { getTestSpeciesList } from '@/lib/providers'
-import { excludedVariants } from '@/lib/utils/excludedSlugs'
-import { getTranslation, TypeKey } from '@/lib/utils/pokeapiHelpers'
+import { excludedVariants } from '@/lib/utils/excluded-slugs'
+import { getTranslation, TypeKey } from '@/lib/utils/pokeapi-helpers'
 import SpeciesCardGrid from '@/components/compounds/SpeciesCardGrid'
 
 export const dynamic = 'force-static'
