+++ components/TypeIcon.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image'
 import clsx from 'clsx/lite'
 import { cva } from 'cva'
-import { TypeLabels, TypeKey } from '@/lib/utils/pokeapiHelpers'
+import { TypeLabels, TypeKey } from '@/lib/utils/pokeapi-helpers'
 
 const variants = cva({
   base: 'flex flex-row items-center justify-center',
