+++ components/compounds/VariantCardSelector.tsx
@@ -4,7 +4,7 @@ import { usePathname, useRouter } from 'next/navigation'
 import { useState } from 'react'
 import clsx from 'clsx/lite'
 import { Radio, RadioGroup } from '@headlessui/react'
-import { Monster } from '@/lib/utils/pokeapiHelpers'
+import { Monster } from '@/lib/utils/pokeapi-helpers'
 import VariantCard from '@/components/compounds/VariantCard'
 
 export default function VariantCardSelector({
