+++ components/details/MonsterHero.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image'
 import clsx from 'clsx/lite'
 import { motion } from 'motion/react'
 import type { Pokemon, PokemonForm, PokemonSpecies } from 'pokedex-promise-v2'
-import { getTranslation } from '@/lib/utils/pokeapiHelpers'
+import { getTranslation } from '@/lib/utils/pokeapi-helpers'
 import TypePill from '@/components/TypePill'
 
 export default function MonsterHero({
