+++ components/shared/FilterBar.tsx
@@ -3,7 +3,7 @@
 import { useMemo } from 'react'
 import clsx from 'clsx/lite'
 import { Listbox, ListboxLabel, ListboxOption } from '@/components/ui/listbox'
-import { DamageClassKey, TypeKey } from '@/lib/utils/pokeapiHelpers'
+import { DamageClassKey, TypeKey } from '@/lib/utils/pokeapi-helpers'
 import DamageClassIcon from '@/components/DamageClassIcon'
 import TypeIcon from '@/components/TypeIcon'
